# docs/swagger.yaml
basePath: /
definitions:
  common.ErrorResponse:
    properties:
      details:
        additionalProperties: true
        type: object
      error_code:
        type: string
      message:
        type: string
    type: object
...
  db.DeletionResponse:
    properties:
      id:
        type: string
    type: object
...
  db.NewQuickLaunchGlobalDefault:
    properties:
      app_id:
        type: string
      quick_launch_id:
        type: string
    type: object
...
  db.QuickLaunchFavorite:
    properties:
      id:
        type: string
      quick_launch_id:
        type: string
      user:
        type: string
    type: object
  db.QuickLaunchGlobalDefault:
    properties:
      app_id:
        type: string
      id:
        type: string
      quick_launch_id:
        type: string
    type: object
...
  db.UpdateQuickLaunchGlobalDefaultRequest:
    properties:
      app_id:
        type: string
      quick_launch_id:
        type: string
    type: object
...
  httphandlers.NewQuickLaunchFavoriteRequest:
    properties:
      quick_launch_id:
        type: string
    type: object
host: localhost:60000
info:
  contact: {}
  description: API for the Discovery Environment's analyses, including Quick Launches
    and settings.
  license:
    name: 3-Clause BSD License
    url: https://github.com/cyverse-de/analyses?tab=License-1-ov-file#readme
  title: Analyses API
  version: "1.0"
paths:
  /quicklaunch/defaults/global:
    get:
      description: Get all of the Quick Launch global defaults that the user has created.
      parameters:
      - description: Username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.QuickLaunchGlobalDefault'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get all Quick Launch global defaults
      tags:
      - quicklaunch-global-defaults
    post:
      consumes:
      - application/json
      description: Add a new Quick Launch global default. Assigns a new UUID.
      parameters:
      - description: Username
        in: query
        name: user
        required: true
        type: string
      - description: Global default to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/db.NewQuickLaunchGlobalDefault'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.QuickLaunchGlobalDefault'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Add a Quick Launch global default
      tags:
      - quicklaunch-global-defaults
  /quicklaunch/defaults/global/{id}:
    delete:
      description: Delete the Quick Launch global default.
      parameters:
      - description: Global Default ID
        in: path
        name: id
        required: true
        type: string
      - description: Username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.DeletionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete a Quick Launch global default
      tags:
      - quicklaunch-global-defaults
    get:
      description: Get a Quick Launch global default.
      parameters:
      - description: Global Default ID
        in: path
        name: id
        required: true
        type: string
      - description: Username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.QuickLaunchGlobalDefault'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get a Quick Launch global default
      tags:
      - quicklaunch-global-defaults
    patch:
      consumes:
      - application/json
      description: Modifies an existing Quick Launch global default.
      parameters:
      - description: Global Default ID
        in: path
        name: id
        required: true
        type: string
      - description: Username
        in: query
        name: user
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/db.UpdateQuickLaunchGlobalDefaultRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.QuickLaunchGlobalDefault'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Update a Quick Launch global default
      tags:
      - quicklaunch-global-defaults
...
  /quicklaunch/favorites:
    get:
      description: Gets all of the user's Quick Launch favorites.
      parameters:
      - description: Username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.QuickLaunchFavorite'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get all Quick Launch favorites
      tags:
      - quicklaunch-favorites
    post:
      consumes:
      - application/json
      description: Adds a favorite Quick Launch to the database for the user. The
        username passed in should already exist. A new UUID will be assigned to the
        favorite and returned with the rest of the record.
      parameters:
      - description: Username
        in: query
        name: user
        required: true
        type: string
      - description: Quick Launch to favorite
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httphandlers.NewQuickLaunchFavoriteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.QuickLaunchFavorite'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Add a Quick Launch favorite
      tags:
      - quicklaunch-favorites
  /quicklaunch/favorites/{id}:
    delete:
      description: Deletes a Quick Launch favorite. Does not delete the actual Quick
        Launch, just the entry that listed it as a favorite for the user.
      parameters:
      - description: Favorite ID
        in: path
        name: id
        required: true
        type: string
      - description: Username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.DeletionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete a Quick Launch favorite
      tags:
      - quicklaunch-favorites
    get:
      description: Gets information about a favorited Quick Launch. Returns a Quick
        Launch UUID which can be passed to the /quicklaunches endpoints to grab more
        information.
      parameters:
      - description: Favorite ID
        in: path
        name: id
        required: true
        type: string
      - description: Username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.QuickLaunchFavorite'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get a Quick Launch favorite
      tags:
      - quicklaunch-favorites
...
swagger: "2.0"